+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [unreleased, v7.0.0-alpha.4]
+
+### Added
+
+- Add extensions to the release
+
+### Changed
+
+- Move extensions documentation from `doc/extensions.md` to `extensions/README.md` for better visibility
+
 ## [v7.0.0-alpha.3] - September 17, 2025
 
 ### Added